backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
mysqlclient==2.2.7
requests==2.32.3
schedule==1.2.2
sqlparse==0.5.3
urllib3==2.3.0
